Move Then-statement from group tests to round tests
Making it available in both
